requirements.txt
# QASPER检索评估系统依赖包
# QASPER Retrieval Evaluation System Dependencies

# 核心科学计算库 / Core scientific computing libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习和深度学习 / Machine learning and deep learning
torch>=1.12.0
transformers>=4.20.0
sentence-transformers>=2.2.0
datasets>=2.0.0
scikit-learn>=1.1.0

# 自然语言处理 / Natural language processing
nltk>=3.7
spacy>=3.4.0
jieba>=0.42.1

# 检索和向量搜索 / Retrieval and vector search
rank-bm25>=0.2.2
faiss-cpu>=1.7.2  # 或使用 faiss-gpu 如果有GPU / or use faiss-gpu if GPU available
rouge-score>=0.1.2

# 数据处理和IO / Data processing and IO
PyYAML>=6.0
tqdm>=4.64.0
psutil>=5.9.0

# 日志和调试 / Logging and debugging
loguru>=0.6.0

# 可选：GPU支持 / Optional: GPU support
# faiss-gpu>=1.7.2  # 取消注释如果使用GPU / Uncomment if using GPU
# torch-audio>=0.12.0  # 如果需要音频处理 / If audio processing needed

# 开发和测试工具 / Development and testing tools
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# 文档生成 / Documentation generation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Jupyter支持 / Jupyter support
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 性能监控 / Performance monitoring
memory-profiler>=0.60.0
line-profiler>=3.5.0

# 实验管理 / Experiment management
wandb>=0.12.0  # 可选：用于实验跟踪 / Optional: for experiment tracking
mlflow>=1.26.0  # 可选：用于实验管理 / Optional: for experiment management

# 网络和HTTP / Network and HTTP
requests>=2.28.0
urllib3>=1.26.0

# 并行处理 / Parallel processing
joblib>=1.1.0
multiprocessing-logging>=0.3.4

# 配置文件处理 / Configuration file processing
configparser>=5.2.0
python-dotenv>=0.20.0

# 时间和日期处理 / Time and date processing
python-dateutil>=2.8.0

# 字符串处理 / String processing
regex>=2022.0.0
unicodedata2>=14.0.0

# 系统监控 / System monitoring
GPUtil>=1.4.0  # 可选：GPU监控 / Optional: GPU monitoring

# 可视化 / Visualization
plotly>=5.0.0  # 可选：交互式可视化 / Optional: interactive visualization

# 注意：根据实际使用情况，某些包可能需要特定版本
# Note: Some packages may require specific versions based on actual usage

# 对于中国用户，可以使用清华源加速下载：
# For Chinese users, you can use Tsinghua mirror for faster download:
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple

# 对于conda用户，建议先安装以下包：
# For conda users, it's recommended to install these packages first:
# conda install numpy pandas scipy scikit-learn
# conda install pytorch torchvision torchaudio pytorch-cuda=11.7 -c pytorch -c nvidia
# conda install -c conda-forge transformers datasets nltk spacy